new commit

+++ wp-content/plugins/popup-maker/classes/Utils/Logging.php
@@ -2,14 +2,15 @@
 /**
  * Debug Logging Utility
  *
- * @package   PUM
- * @copyright Copyright (c) 2023, Code Atlantic LLC
+ * @package   PopupMaker
+ * @copyright Copyright (c) 2024, Code Atlantic LLC
  */
 
 /**
  * Class PUM_Utils_Logging
  *
  * @since 1.8.0
+ * @deprecated 1.21.0 Use \PopupMaker\logging() instead.
  */
 class PUM_Utils_Logging {
 
@@ -37,14 +38,14 @@ class PUM_Utils_Logging {
 	/**
 	 * File system API.
 	 *
-	 * @var WP_Filesystem_Base
+	 * @var WP_Filesystem_Base|null
 	 */
 	private $fs;
 
 	/**
 	 * Log file content.
 	 *
-	 * @var string
+	 * @var string|null
 	 */
 	private $content;
 
@@ -58,43 +59,85 @@ class PUM_Utils_Logging {
 	/**
 	 * Get instance.
 	 *
-	 * @return PUM_Utils_Logging
+	 * @deprecated 1.21.0 Use \PopupMaker\plugin()->get( 'logging' ) instead.
+	 *
+	 * @return \PopupMaker\Services\Logging
 	 */
 	public static function instance() {
-		if ( ! isset( self::$instance ) ) {
-			self::$instance = new self();
-		}
-
-		return self::$instance;
+		return \PopupMaker\plugin()->get( 'logging' );
 	}
 
 	/**
 	 * Get things started
 	 */
 	public function __construct() {
-		// On shutdown, save the log file.
-		add_action( 'shutdown', [ $this, 'save_logs' ] );
-
-		$this->fs = $this->file_system();
-
-		$this->is_writable = false !== $this->fs && 'direct' === $this->fs->method;
-
-		$upload_dir = PUM_Helpers::get_upload_dir();
-		if ( ! $this->fs || ! $this->fs->is_writable( $upload_dir['basedir'] ) ) {
-			$this->is_writable = false;
+		if ( $this->disabled() ) {
+			return;
 		}
 
+		$this->init_fs();
 		$this->init();
+
+		// On shutdown, save the log file.
+		add_action( 'shutdown', [ $this, 'save_logs' ] );
 	}
 
 	/**
-	 * Check if logging is enabled.
+	 * Check if logging is disabled.
+	 *
+	 * @deprecated 1.21.0 Use \PopupMaker\logging()->disabled() instead.
+	 *
+	 * @return bool
+	 */
+	public function disabled() {
+		return \PopupMaker\logging()->disabled();
+	}
+
+	/**
+	 * Check if logging is writeable & not disabled.
+	 *
+	 * If this is true the $fs property will be set.
 	 *
 	 * @return bool
 	 */
 	public function enabled() {
-		// Disable logging by adding define( 'PUM_DISABLE_LOGGING', true );.
-		return $this->is_writable && ( ! defined( 'PUM_DISABLE_LOGGING' ) || ! PUM_DISABLE_LOGGING );
+		return ! $this->disabled() && $this->is_writable;
+	}
+
+	/**
+	 * Initialize the file system.
+	 *
+	 * - Check if the file system is writable.
+	 * - Check if the upload directory is writable.
+	 * - Set the file system instance.
+	 *
+	 * @return void
+	 */
+	public function init_fs() {
+		$fs       = $this->file_system();
+		$this->fs = false !== $fs ? $fs : null;
+
+		// If the file system is not set, we can't check if it's writable.
+		if ( null === $this->fs ) {
+			return;
+		}
+
+		$this->is_writable = 'direct' === $this->fs->method;
+
+		if ( ! $this->is_writable ) {
+			return;
+		}
+
+		$upload_dir = PUM_Helpers::get_upload_dir();
+
+		if ( false === $upload_dir ) {
+			$this->is_writable = false;
+			return;
+		}
+
+		if ( ! $this->fs->is_writable( $upload_dir['basedir'] ) ) {
+			$this->is_writable = false;
+		}
 	}
 
 	/**
@@ -120,14 +163,30 @@ class PUM_Utils_Logging {
 	}
 
 	/**
-	 * Get things started
+	 * Get things started.
+	 *
+	 * - Get filetoken & name.
+	 * - Check if old log file exists, move it to new location.
+	 * - Check if new log file exists, if not create it.
+	 * - Set log content.
+	 * - Truncate long log files.
+	 *
+	 * @return void
 	 */
 	public function init() {
-		$upload_dir = PUM_Helpers::get_upload_dir();
+		if ( ! $this->enabled() || null === $this->fs ) {
+			return;
+		}
+
+		$upload_dir = \PopupMaker\get_upload_dir();
+
+		if ( false === $upload_dir ) {
+			return;
+		}
 
 		$file_token = get_option( 'pum_debug_log_token' );
 		if ( false === $file_token ) {
-			$file_token = uniqid( wp_rand(), true );
+			$file_token = uniqid( (string) wp_rand(), true );
 			update_option( 'pum_debug_log_token', $file_token );
 		}
 
@@ -164,17 +223,18 @@ class PUM_Utils_Logging {
 	/**
 	 * Retrieves the url to the file
 	 *
-	 * @returns string
+	 * @return string
 	 * @since 1.12.0
 	 */
 	public function get_file_url() {
-		return PUM_Helpers::get_upload_dir_url( $this->filename );
+		$url = \PopupMaker\get_upload_dir_url( $this->filename );
+		return is_string( $url ) ? $url : '';
 	}
 
 	/**
 	 * Retrieve the log data
 	 *
-	 * @return string
+	 * @return string|null
 	 */
 	public function get_log() {
 		return $this->get_log_content();
@@ -184,15 +244,17 @@ class PUM_Utils_Logging {
 	 * Log message to file
 	 *
 	 * @param string $message The message to log.
+	 * @return void
 	 */
 	public function log( $message = '' ) {
-		$this->write_to_log( wp_date( 'Y-n-d H:i:s' ) . ' - ' . $message );
+		$this->write_to_log( ( function_exists( 'wp_date' ) ? wp_date( 'Y-n-d H:i:s' ) : date( 'Y-n-d H:i:s', current_time( 'timestamp' ) ) ) . ' - ' . $message );
 	}
 
 	/**
 	 * Log unique message to file.
 	 *
 	 * @param string $message The unique message to log.
+	 * @return void
 	 */
 	public function log_unique( $message = '' ) {
 		$contents = $this->get_log_content();
@@ -207,7 +269,7 @@ class PUM_Utils_Logging {
 	/**
 	 * Get the log file contents.
 	 *
-	 * @return string
+	 * @return string|null
 	 */
 	public function get_log_content() {
 		if ( ! isset( $this->content ) ) {
@@ -220,8 +282,8 @@ class PUM_Utils_Logging {
 	/**
 	 * Set the log file contents in memory.
 	 *
-	 * @param mixed $content The content to set.
-	 * @param bool  $save    Whether to save the content to the file immediately.
+	 * @param string $content The content to set.
+	 * @param bool   $save    Whether to save the content to the file immediately.
 	 * @return void
 	 */
 	private function set_log_content( $content, $save = false ) {
@@ -235,21 +297,21 @@ class PUM_Utils_Logging {
 	/**
 	 * Retrieve the contents of a file.
 	 *
-	 * @param string|boolean $file File to get contents of.
-	 *
-	 * @return string
+	 * @param string|false $file File path to get contents of, or false to use default log file.
+	 * @return string File contents or empty string on failure.
 	 */
 	protected function get_file( $file = false ) {
 		$file = $file ? $file : $this->file;
 
-		if ( ! $this->enabled() ) {
+		if ( ! $this->enabled() || null === $this->fs ) {
 			return '';
 		}
 
 		$content = '';
 
 		if ( $this->fs->exists( $file ) ) {
-			$content = $this->fs->get_contents( $file );
+			$file_content = $this->fs->get_contents( $file );
+			$content      = is_string( $file_content ) ? $file_content : '';
 		}
 
 		return $content;
@@ -259,6 +321,7 @@ class PUM_Utils_Logging {
 	 * Write the log message
 	 *
 	 * @param string $message The message to write.
+	 * @return void
 	 */
 	protected function write_to_log( $message = '' ) {
 		if ( ! $this->enabled() ) {
@@ -277,13 +340,15 @@ class PUM_Utils_Logging {
 
 	/**
 	 * Save the current contents to file.
+	 *
+	 * @return void
 	 */
 	public function save_logs() {
-		if ( ! $this->enabled() ) {
+		if ( ! $this->enabled() || null === $this->fs ) {
 			return;
 		}
 
-		$this->fs->put_contents( $this->file, $this->content, FS_CHMOD_FILE );
+		$this->fs->put_contents( $this->file, $this->content ?? '', FS_CHMOD_FILE );
 	}
 
 	/**
@@ -293,18 +358,20 @@ class PUM_Utils_Logging {
 	 */
 	public function count_lines() {
 		$file  = $this->get_log_content();
-		$lines = explode( "\r\n", $file );
+		$lines = explode( "\r\n", $file ?? '' );
 
 		return count( $lines );
 	}
 
 	/**
 	 * Truncates a log file to maximum of 250 lines.
+	 *
+	 * @return void
 	 */
 	public function truncate_log() {
 		$content           = $this->get_log_content();
-		$lines             = explode( "\r\n", $content );
-		$lines             = array_slice( $lines, 0, 250 ); // 50 is how many lines you want to keep
+		$lines             = explode( "\r\n", $content ?? '' );
+		$lines             = array_slice( $lines, 0, 250 ); // 250 is how many lines you want to keep
 		$truncated_content = implode( "\r\n", $lines );
 		$this->set_log_content( $truncated_content, true );
 	}
@@ -315,15 +382,20 @@ class PUM_Utils_Logging {
 	 * @return void
 	 */
 	public function setup_new_log() {
-		$this->set_log_content( "Popup Maker Debug Logs:\r\n" . wp_date( 'Y-n-d H:i:s' ) . " - Log file initialized\r\n", true );
+		$this->set_log_content( "Popup Maker Debug Logs:\r\n" . ( function_exists( 'wp_date' ) ? wp_date( 'Y-n-d H:i:s' ) : date( 'Y-n-d H:i:s', current_time( 'timestamp' ) ) ) . " - Log file initialized\r\n", true );
 	}
 
 	/**
 	 * Delete the log file.
+	 *
+	 * @return void
 	 */
 	public function clear_log() {
 		// Delete the file.
-		@$this->fs->delete( $this->file );
+		if ( null !== $this->fs ) {
+			// phpcs:ignore WordPress.PHP.NoSilencedErrors.Discouraged
+			@$this->fs->delete( $this->file );
+		}
 
 		if ( $this->enabled() ) {
 			$this->setup_new_log();
@@ -333,9 +405,10 @@ class PUM_Utils_Logging {
 	/**
 	 * Log a deprecated notice.
 	 *
-	 * @param string $func_name Function name.
-	 * @param string $version Versoin deprecated.
-	 * @param string $replacement Replacement function (optional).
+	 * @param non-empty-string $func_name   Function name.
+	 * @param non-empty-string $version     Version deprecated.
+	 * @param string|null      $replacement Replacement function (optional).
+	 * @return void
 	 */
 	public function log_deprecated_notice( $func_name, $version, $replacement = null ) {
 		if ( ! is_null( $replacement ) ) {
