+++ wp-content/plugins/simply-static/src/tasks/class-ss-setup-task.php
@@ -33,34 +33,27 @@ class Setup_Task extends Task {
 			Util::debug_log( 'Creating archive directory: ' . $archive_dir );
 			$create_dir = wp_mkdir_p( $archive_dir );
 			if ( $create_dir === false ) {
-				return new \WP_Error( 'cannot_create_archive_dir', sprintf(__('Cannot create archive directory %s'), $archive_dir) );
+				return new \WP_Error( 'cannot_create_archive_dir', sprintf( __( 'Cannot create archive directory %s' ), $archive_dir ) );
 			}
 		}
 
-		// TODO: Add a way for the user to perform this, optionally, so that we
-		// don't need to do it every time. Then enable the two commented-out
-		// sections below.
 		$use_single = get_option( 'simply-static-use-single' );
 		$use_build  = get_option( 'simply-static-use-build' );
+		$type       = $this->options->get( 'generate_type' );
 
-		if ( empty( $use_build ) && empty( $use_single ) ) {
-			Page::query()->delete_all();
+		if ( ! $type ) {
+			$type = 'export';
 		}
 
-		// clear out any saved error messages on pages
-		//Page::query()
-		//->update_all( 'error_message', null );
-
-		// delete pages that we can't process
-		//Page::query()
-		//->where( 'http_status_code IS NULL OR http_status_code NOT IN (?)', implode( ',', Page::$processable_status_codes ) )
-		//->delete_all();
+		if ( empty( $use_build ) && empty( $use_single ) && 'export' === $type ) {
+			Page::query()->delete_all();
+		}
 
 		// add origin url and additional urls/files to database.
 		$additional_urls = apply_filters( 'ss_setup_task_additional_urls', $this->options->get( 'additional_urls' ) );
 
-		self::add_origin_and_additional_urls_to_db( $additional_urls );
-		self::add_additional_files_to_db( $this->options->get( 'additional_files' ) );
+		$this->add_origin_and_additional_urls_to_db( $additional_urls );
+		$this->add_additional_files_to_db( $this->options->get( 'additional_files' ) );
 
 		do_action( 'ss_after_setup_task' );
 
@@ -74,7 +67,7 @@ class Setup_Task extends Task {
 	 *
 	 * @return void
 	 */
-	public static function add_origin_and_additional_urls_to_db( $additional_urls ) {
+	public function add_origin_and_additional_urls_to_db( $additional_urls ) {
 		$origin_url = trailingslashit( Util::origin_url() );
 		Util::debug_log( 'Adding origin URL to queue: ' . $origin_url );
 		$static_page = Page::query()->find_or_initialize_by( 'url', $origin_url );
@@ -82,28 +75,34 @@ class Setup_Task extends Task {
 		$static_page->found_on_id = 0;
 		$static_page->save();
 
-		$urls = array_unique( Util::string_to_array( $additional_urls ) );
+		// Convert raw textarea to array but avoid array_unique on the full dataset to reduce peak memory.
+		$raw_urls = Util::string_to_array( $additional_urls );
+		$urls     = apply_filters( 'ss_additional_urls', $raw_urls );
 
-		// Add additional URls for SimplyCDN integration.
-		$options = get_option( 'simply-static' );
-
-		if ( isset( $options['ssh_404_page_id'] ) ) {
-			$urls[] = get_permalink( $options['ssh_404_page_id'] );
+		// Process Additional URLs in batches to avoid memory exhaustion with very large lists.
+		$batch_size = (int) apply_filters( 'ss_additional_urls_batch_size', 50 );
+		if ( $batch_size < 1 ) {
+			$batch_size = 50;
 		}
 
-		if ( isset( $options['ssh_thank_you_page_id'] ) ) {
-			$urls[] = get_permalink( $options['ssh_thank_you_page_id'] );
-		}
+		foreach ( array_chunk( (array) $urls, $batch_size ) as $chunk ) {
+			// De-duplicate within each chunk to minimize duplicate operations without loading entire list into memory.
+			$chunk = array_unique( $chunk );
 
-		$urls = apply_filters( 'ss_additional_urls', $urls );
+			foreach ( $chunk as $url ) {
+				if ( Util::is_local_url( $url ) ) {
+					Util::debug_log( 'Adding additional URL to queue: ' . $url );
+					$static_page = Page::query()->find_or_initialize_by( 'url', $url );
+					$static_page->set_status_message( __( 'Additional URL', 'simply-static' ) );
+					$static_page->found_on_id = 0;
+					$static_page->save();
+				}
+			}
 
-		foreach ( $urls as $url ) {
-			if ( Util::is_local_url( $url ) ) {
-				Util::debug_log( 'Adding additional URL to queue: ' . $url );
-				$static_page = Page::query()->find_or_initialize_by( 'url', $url );
-				$static_page->set_status_message( __( "Additional URL", 'simply-static' ) );
-				$static_page->found_on_id = 0;
-				$static_page->save();
+			// Free memory for processed chunk explicitly.
+			unset( $chunk );
+			if ( function_exists( 'gc_collect_cycles' ) ) {
+				gc_collect_cycles();
 			}
 		}
 	}
@@ -115,47 +114,98 @@ class Setup_Task extends Task {
 	 *
 	 * @return void
 	 */
-	public static function add_additional_files_to_db( $additional_files ) {
-
+	public function add_additional_files_to_db( $additional_files ) {
 		$additional_files = apply_filters( 'ss_additional_files', Util::string_to_array( $additional_files ) );
 
-		// Add additional files for SimplyCDN integration.
-		$options = get_option( 'simply-static' );
+		// Add robots.txt if exists.
+		$robots_txt = ABSPATH . 'robots.txt';
 
-		if ( isset( $options['ssh_use_forms'] ) ) {
-			$additional_files[] = SIMPLY_STATIC_PATH . '/src/integrations/simply-cdn/assets/ssh-form-webhook.js';
+		if ( file_exists( $robots_txt ) ) {
+			$additional_files[] = $robots_txt;
 		}
 
-		// Convert additional files to URLs and add to queue.
-		foreach ( $additional_files as $item ) {
-			// If item is a file, convert to url and insert into database.
-			// If item is a directory, recursively iterate and grab all files,
-			// and for each file, convert to url and insert into database.
-			if ( file_exists( $item ) ) {
-				if ( is_file( $item ) ) {
-					$url = self::convert_path_to_url( $item );
-					Util::debug_log( "File " . $item . ' exists; adding to queue as: ' . $url );
-					$static_page = Page::query()
-					                   ->find_or_create_by( 'url', $url );
-					$static_page->set_status_message( __( "Additional File", 'simply-static' ) );
-					// setting found_on_id to 0 since this was user-specified
-					$static_page->found_on_id = 0;
-					$static_page->save();
-				} else {
-					Util::debug_log( "Adding files from directory: " . $item );
-					$iterator = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $item, \RecursiveDirectoryIterator::SKIP_DOTS ) );
+		// Add llms.txt if exists.
+		$llms_txt = ABSPATH . 'llms.txt';
 
-					foreach ( $iterator as $file_name => $file_object ) {
-						$url = self::convert_path_to_url( $file_name );
-						Util::debug_log( "Adding file " . $file_name . ' to queue as: ' . $url );
-						$static_page = Page::query()->find_or_initialize_by( 'url', $url );
-						$static_page->set_status_message( __( "Additional Dir", 'simply-static' ) );
+		if ( file_exists( $llms_txt ) ) {
+			$additional_files[] = $llms_txt;
+		}
+
+		// Add feeds if enabled.
+		if ( $this->options->get( 'add_feeds' ) ) {
+			// Create feed directory it doesn't exist.
+			$feed_directory = untrailingslashit( $this->options->get_archive_dir() ) . '/feed';
+
+			if ( ! file_exists( $feed_directory ) ) {
+				wp_mkdir_p( $feed_directory );
+			}
+
+			// Create index.html file for feed directory.
+			file_put_contents( $feed_directory . '/index.html',
+				'<!DOCTYPE html>
+			<html>
+				<head>
+					<title>Redirecting...</title>
+					<meta http-equiv="refresh" content="0;url=index.xml">
+				</head>
+				<body>
+					<script type="text/javascript">
+						window.location = "index.xml";
+					</script>
+					<p>You are being redirected to <a href="index.xml">index.xml</a></p>
+				</body>
+			</html>'
+			);
+
+			// Add feed redirect file to additional files.
+			$additional_files[] = $feed_directory . '/index.html';
+		}
+
+		// Process Additional Files/Directories in batches to reduce peak memory usage.
+		$batch_size = (int) apply_filters( 'ss_additional_files_batch_size', 50 );
+		if ( $batch_size < 1 ) {
+			$batch_size = 50;
+		}
+
+		foreach ( array_chunk( (array) $additional_files, $batch_size ) as $chunk ) {
+			foreach ( $chunk as $item ) {
+				// If item is a file, convert to url and insert into database.
+				// If item is a directory, recursively iterate and grab all files,
+				// and for each file, convert to url and insert into database.
+				if ( file_exists( $item ) ) {
+					if ( is_file( $item ) ) {
+						$url = self::convert_path_to_url( $item );
+						Util::debug_log( "File " . $item . ' exists; adding to queue as: ' . $url );
+						$static_page = Page::query()
+						                   ->find_or_create_by( 'url', $url );
+						$static_page->set_status_message( __( "Additional File", 'simply-static' ) );
+						// setting found_on_id to 0 since this was user-specified
 						$static_page->found_on_id = 0;
+						$static_page->handler     = Additional_File_Handler::class;
 						$static_page->save();
+					} else {
+						Util::debug_log( "Adding files from directory: " . $item );
+						$iterator = new \RecursiveIteratorIterator( new \RecursiveDirectoryIterator( $item, \RecursiveDirectoryIterator::SKIP_DOTS ) );
+
+						foreach ( $iterator as $file_name => $file_object ) {
+							$url = self::convert_path_to_url( $file_name );
+							Util::debug_log( "Adding file " . $file_name . ' to queue as: ' . $url );
+							$static_page = Page::query()->find_or_initialize_by( 'url', $url );
+							$static_page->set_status_message( __( "Additional Dir", 'simply-static' ) );
+							$static_page->handler     = Additional_File_Handler::class;
+							$static_page->found_on_id = 0;
+							$static_page->save();
+						}
 					}
+				} else {
+					Util::debug_log( "File doesn't exist: " . $item );
 				}
-			} else {
-				Util::debug_log( "File doesn't exist: " . $item );
+			}
+
+			// Free memory for processed chunk explicitly.
+			unset( $chunk );
+			if ( function_exists( 'gc_collect_cycles' ) ) {
+				gc_collect_cycles();
 			}
 		}
 	}
@@ -189,10 +239,20 @@ class Setup_Task extends Task {
 	 * @return true|\WP_Error True on success, WP_Error otherwise.
 	 */
 	public function delete_temp_static_files() {
-		$options = Options::instance();
-		$dir     = $options->get( 'temp_files_dir' );
+		$options           = Options::instance();
+		$dir               = $options->get( 'temp_files_dir' );
+		$delete_temp_files = apply_filters( 'ss_delete_temp_files', true );
 
-		if ( false === file_exists( $dir ) ) {
+		if ( ! $delete_temp_files ) {
+			return false;
+		}
+
+		if ( empty( $dir ) ) {
+			$upload_dir = wp_upload_dir();
+			$dir        = $upload_dir['basedir'] . DIRECTORY_SEPARATOR . 'simply-static' . DIRECTORY_SEPARATOR . 'temp-files';
+		}
+
+		if ( false === file_exists( $dir ) || 'update' === $options->get( 'generate_type' ) ) {
 			return false;
 		}
